fix: resolve align issues in front page
separate styles in a separate file

create a typescript interface for data
